Extract `canCurrentUserEditPost` to a separate selector
